Rename ActiveOffersTable => ActiveLoansTable
