+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/ActiveLoansTable.module.less
