+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/ActiveLoansTable.tsx
@@ -6,17 +6,17 @@ import Table from '@banx/components/Table'
 import { Loan } from '@banx/api/core'
 import { isLoanLiquidated, isLoanTerminating } from '@banx/utils'
 
-import { EMPTY_MESSAGE } from '../../constants'
+import { EMPTY_LOANS_MESSAGE } from '../../constants'
 import { getTableColumns } from './columns'
 import { useLenderLoansAndOffers } from './hooks'
 
-import styles from './ActiveOffersTable.module.less'
+import styles from './ActiveLoansTable.module.less'
 
-interface ActiveOffersTableProps {
+interface ActiveLoansTableProps {
   loans: Loan[]
 }
 
-const ActiveOffersTable: FC<ActiveOffersTableProps> = ({ loans }) => {
+const ActiveLoansTable: FC<ActiveLoansTableProps> = ({ loans }) => {
   const { offers, updateOrAddLoan, updateOrAddOffer } = useLenderLoansAndOffers()
 
   const columns = getTableColumns({ offers, updateOrAddOffer, updateOrAddLoan })
@@ -38,7 +38,7 @@ const ActiveOffersTable: FC<ActiveOffersTableProps> = ({ loans }) => {
     }
   }, [])
 
-  if (!loans.length) return <EmptyList message={EMPTY_MESSAGE} />
+  if (!loans.length) return <EmptyList message={EMPTY_LOANS_MESSAGE} />
 
   return (
     <Table
@@ -51,4 +51,4 @@ const ActiveOffersTable: FC<ActiveOffersTableProps> = ({ loans }) => {
   )
 }
 
-export default ActiveOffersTable
+export default ActiveLoansTable
