+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/index.ts
@@ -1,1 +1,1 @@
-export { default } from './ActiveOffersTable'
+export { default } from './ActiveLoansTable'
