+++ src/pages/OffersPage/components/OffersTabContent/components/OfferCard/OfferCard.module.less
@@ -28,6 +28,7 @@
 .mainOfferInfo {
   display: flex;
   flex-direction: column;
+  min-width: 300px;
   gap: 10px;
 }
 .mainOfferStats {
